+++ src/scripts/Tasks/tasksSingleCompontentBuilder.js
